Split the two files my changes pushed over the size guard
CI failed on dispatch/tests.rs (381) and dispatch.rs (301), neither of
which had a transition exception to lean on.

`handle_tools_call` moves to its own module. It is a real seam rather than
a convenient cut: everything in it is about the MCP tool-call envelope --
unwrapping `arguments`, routing the stateful pcap job tools, deciding
whether a failure is a tool result or a protocol fault. What stays in
dispatch.rs is JSON-RPC method routing.

The tests split the same way. `tests.rs` keeps protocol lifecycle;
`policy_tests.rs` takes the target policy, error shape and param tolerance
added by this branch.
